refactor: základní rozdělení logiky do souborů
